requirements.txt
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
pyyaml>=6.0
pokerkit>=0.5.0
pytest>=7.4.0
